+++ distro/config.go
@@ -18,17 +18,33 @@ import (
 	"fmt"
 	"net/url"
 	"os"
+	"strings"
+
+	"go.opentelemetry.io/contrib/propagators/aws/xray"
+	"go.opentelemetry.io/contrib/propagators/b3"
+	"go.opentelemetry.io/contrib/propagators/jaeger"
+	"go.opentelemetry.io/contrib/propagators/ot"
+	"go.opentelemetry.io/otel/propagation"
 )
 
 // Environment variable keys that set values of the configuration.
 const (
+	// Access token added to exported data.
 	accessTokenKey = "SPLUNK_ACCESS_TOKEN"
+
+	// OpenTelemetry TextMapPropagator to set as global.
+	otelPropagatorsKey = "OTEL_PROPAGATORS"
+
+	// FIXME: support OTEL_SPAN_LINK_COUNT_LIMIT
+	// FIXME: support OTEL_ATTRIBUTE_VALUE_LENGTH_LIMIT
+	// FIXME: support OTEL_TRACES_EXPORTER
 )
 
 // config is the configuration used to create and operate an SDK.
 type config struct {
 	AccessToken string
 	Endpoint    string
+	Propagator  propagation.TextMapPropagator
 }
 
 // newConfig returns a validated config with Splunk defaults.
@@ -40,6 +56,14 @@ func newConfig(opts ...Option) (*config, error) {
 	for _, o := range opts {
 		o.apply(c)
 	}
+
+	// Apply default field values if they were not set.
+	if c.Propagator == nil {
+		c.Propagator = loadPropagator(
+			envOr(otelPropagatorsKey, "tracecontext,baggage"),
+		)
+	}
+
 	if err := c.Validate(); err != nil {
 		return nil, err
 	}
@@ -47,7 +71,7 @@ func newConfig(opts ...Option) (*config, error) {
 }
 
 // Validate ensures c is valid, otherwise returning an appropriate error.
-func (c config) Validate() error {
+func (c *config) Validate() error {
 	var errs []string
 
 	if c.Endpoint != "" {
@@ -60,6 +84,82 @@ func (c config) Validate() error {
 		return fmt.Errorf("invalid config: %v", errs)
 	}
 	return nil
+}
+
+type nonePropagatorType struct{ propagation.TextMapPropagator }
+
+// nonePropagator signals the disablement of setting a global
+// TextMapPropagator.
+var nonePropagator = nonePropagatorType{}
+
+// propagators maps environment variable values to TextMapPropagator creation
+// functions.
+var propagators = map[string]func() propagation.TextMapPropagator{
+	// W3C Trace Context.
+	"tracecontext": func() propagation.TextMapPropagator {
+		return propagation.TraceContext{}
+	},
+	// W3C Baggage
+	"baggage": func() propagation.TextMapPropagator {
+		return propagation.Baggage{}
+	},
+	// B3 Single
+	"b3": func() propagation.TextMapPropagator {
+		return b3.New(b3.WithInjectEncoding(b3.B3SingleHeader))
+	},
+	// B3 Multi
+	"b3multi": func() propagation.TextMapPropagator {
+		return b3.New(b3.WithInjectEncoding(b3.B3MultipleHeader))
+	},
+	// Jaeger
+	"jaeger": func() propagation.TextMapPropagator {
+		return jaeger.Jaeger{}
+	},
+	// AWS X-Ray.
+	"xray": func() propagation.TextMapPropagator {
+		return xray.Propagator{}
+	},
+	// OpenTracing Trace
+	"ottrace": func() propagation.TextMapPropagator {
+		return ot.OT{}
+	},
+	// None, explicitly do not set a global propagator.
+	"none": func() propagation.TextMapPropagator {
+		return nonePropagator
+	},
+}
+
+func loadPropagator(name string) propagation.TextMapPropagator {
+	var props []propagation.TextMapPropagator
+	for _, part := range strings.Split(name, ",") {
+		factory, ok := propagators[part]
+		if !ok {
+			// Skip invalid data.
+			// TODO: log this.
+			continue
+		}
+
+		p := factory()
+		if p == nonePropagator {
+			// Make sure the disablement of the global propagator does not get
+			// lost as a composite below.
+			return p
+		}
+		props = append(props, p)
+	}
+
+	switch len(props) {
+	case 0:
+		// Default to "tracecontext,baggage".
+		return propagation.NewCompositeTextMapPropagator(
+			propagation.TraceContext{},
+			propagation.Baggage{},
+		)
+	case 1:
+		return props[0]
+	default:
+		return propagation.NewCompositeTextMapPropagator(props...)
+	}
 }
 
 // envOr returns the environment variable value associated with key if it
@@ -84,19 +184,45 @@ func (fn optionFunc) apply(c *config) {
 	fn(c)
 }
 
-// WithEndpoint configures the endpoint telemetry is sent to.
-// Setting empty string results in no operation.
+// WithEndpoint configures the endpoint telemetry is sent to. Passing an empty
+// string results in the default value being used.
 func WithEndpoint(endpoint string) Option {
 	return optionFunc(func(c *config) {
 		c.Endpoint = endpoint
 	})
 }
 
-// WithAccessToken configures the authentication token
-// allowing exporters to send data directly to a Splunk back-end.
-// Setting empty string results in no operation.
+// WithAccessToken configures the authentication token used to authenticate
+// telemetry sent directly to Splunk Observability Cloud. Passing an empty
+// string results in no authentication token being used, and assumes
+// authentication is handled by another system.
+//
+// The SPLUNK_ACCESS_TOKEN environment variable value is used if this Option
+// is not provided.
 func WithAccessToken(accessToken string) Option {
 	return optionFunc(func(c *config) {
 		c.AccessToken = accessToken
+	})
+}
+
+// WithPropagator configures the OpenTelemetry TextMapPropagator set as the
+// global TextMapPropagator. Passing nil will prevent any TextMapPropagator
+// from being set.
+//
+// The OTEL_PROPAGATORS environment variable value is used if this Option is
+// not provided.
+//
+// By default, a tracecontext and baggage TextMapPropagator is set as the
+// global TextMapPropagator if this is not provided or the OTEL_PROPAGATORS
+// environment variable is not set.
+func WithPropagator(p propagation.TextMapPropagator) Option {
+	return optionFunc(func(c *config) {
+		if p == nil {
+			// Set to nonePropagator so when environment variable overrides
+			// are applied this is distinguishable from no WithPropagator
+			// option being passed.
+			p = nonePropagator
+		}
+		c.Propagator = p
 	})
 }